# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Registro Elettronico V2 API
  description: API REST per la gestione scolastica (Auth, Utenti, Voti, Presenze, Comunicazioni, Documenti, Colloqui, Agenda).
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
paths:
...
  /auth/login:
    post:
      summary: Autenticazione utente
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login effettuato con successo (restituisce JWT access_token)
        '401':
          description: Credenziali non valide
...
  /auth/register:
    post:
      summary: Registrazione nuovo utente (Richiede JWT con ruoli autorizzati)
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Utente creato
        '403':
          description: Permessi insufficienti
...
  /auth/mfa/setup:
    post:
      summary: Inizializza setup TOTP 2FA
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Secret TOTP e QR code URI
  /auth/mfa/verify:
    post:
      summary: Verifica codice 2FA e attiva MFA
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: MFA attivato
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT